+++ src/index.js
@@ -1,2 +1,3 @@
 export { default as default } from './Form';
+export { default as Form } from './Form';
 export { default as Errors } from './Errors';